Refactored the code for Sexp_call, and made minor edits and fixes to Sexp_rcall.
Now Sexp_call, exposed as SexpClosure.__call__(), is merely a wrapper for SexpClosure.rcall().

This has a cost for functions with large number of arguments, but has the benefit of simplifying
the code base. That last aspect was discussed on the mailing-list with Nathaniel Smith
- and I initially failed to see that __call__() and rcall() could remain different, as the cost
is most certainly limited, and restricted to few use cases -.

--HG--
branch : version_2.1.x