Forbid double activation of a converter + add tests
numpy2ri and pandas2ri save the previous conversion methods (py2ri,
ri2ro, and py2ro) in global variables. This is done for two reasons:
 * the new converter may delegate some conversions to the old one
 * the old converter is restored when deactivate() is called.

If a module is activated twice, the old converter backup is overwritten
and cannot be restored anymore, moreover the delegate conversions fail.

Added tests to testNumpyConversions and testPandasConversions to catch
this.

--HG--
branch : fix-activate-deactivate