Allow optional building and use API-mode.
cffi can be used in ABI and API mode. This allows the building
of an API-mode interface (requires compilation), and an optional
use of that interface. The choice is made through an environment
variable (`RPY2_CFFI_MODE`).

At the moment starting the embedded R in API mode makes R unable
to find its own packages. The root of the issue is not yet known.