Better handling of the various build situations.
Combinations of ABI/API/BOTH/ANY along with the capabilities
of the system on which the package is built should work better.
The summary message at the end of building or installing, telling
which of the ABI and API modes were built/installed should also
be more accurate.

There is also a stub of DLL opening for Windows. It is broken though.