enh: look for R path in HKEY_CURRENT_USER (Windows registry) (#788)
On Windows, R can be installed by a user (not admin) in which case the binary is installed in the user's home directory. In addition, the R installer edits the registry in "HKEY_CURRENT_USER" instead of in "HKEY_LOCAL_MACHINE". This PR makes rpy2 also look in "HKEY_CURRENT_USER".