Fix unit test for interrupting R.
There are two issues in the existing test:
- The python source code is indented, fixed by using textwrap.dedent().
- It seems to change console writing using an old method; Not sure how
  critical this is to implement, but I've mirrored the code to do this
  from test_set_consolewrite_print.

I also make the exception in the test more specific, an RRuntimeError,
which is what I see when testing the feature locally.

While not critical to fixing the test, I tried to clean up the code used
to check the child process return code by swapping time.sleep() and
child_proc.poll() out for child_proc.wait(). This should achieve what
the old code did, but may let the test complete earlier if the child
responds quickly enough to the SIGINT. For test failures, this will
wind up throwing a TimeoutError (instead of an assertion failure with
ret_code=None as it does before this change).